Adds message table columns and model relations with User

+++ database/migrations/2018_12_15_033819_add_default_user_account.php
@@ -31,6 +31,9 @@ class AddDefaultUserAccount extends Migration
     {
         $user = \App\User::where('email', env("DEFAULT_EMAIL", "email@example.com"))->first();
 
-        $user->delete();
+        if (!is_null($user)) {
+            $user->delete();
+        }
+
     }
 }
